+++ include/H2DE/H2DE.h
@@ -1,5 +1,4 @@
 /**
- * Version: 2.0.19
- */
+ * Version: 2.0.20 */
 
 #include <H2DE_engine.h>